requirements.txt
Django==2.0
djangorestframework==3.7.4
pytz==2017.3
